requirements.txt
asgiref==3.4.1
boto3==1.19.1
botocore==1.22.1
certifi==2021.5.30
dj-database-url==0.5.0
Django==3.2.7
django-admin-interface==0.17.2
django-colorfield==0.4.4
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-heroku==0.3.1
django-storages==1.12.2
django-summernote==0.8.11.6
gunicorn==20.1.0
jmespath==0.10.0
Pillow==8.3.2
psycopg2==2.9.1
python-dateutil==2.8.2
pytz==2021.1
s3transfer==0.5.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.7
whitenoise==5.3.0
